+++ pkg/strace/internal/abi/abi.go
