+++ pkg/strace/internal/abi/abi_linux.go
