+++ pkg/strace/syscall_linux.go
@@ -22,7 +22,7 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/u-root/u-root/pkg/abi"
+	"github.com/u-root/u-root/pkg/strace/internal/abi"
 	"golang.org/x/sys/unix"
 )
